dhcpv4: export two functions
And add dummy functions for when DHCPv4 isn't compiled in. At the same
time, move all DHCPv4 declarations to the dhcpv4 header file.
